Revise tests to show failures on Wagtail 6
The old tests relied on the assumption that page.save_revision() was an accurate representation of what happens when saving a draft edit in the admin. On Wagtail 6, saving a draft also updates the real database record if the page is not live, meaning that it is no longer a stable place to stash the 'approved' revision as it was when the experiment was activated.